Fix current animation to fallback to idle on top-down mode
If the entity is eg. a player, performace is dynamic, so the first "if" applies. If on top-down mode, we'll always go to
```javascript
else if (_c.TOP_DOWN) {
```
and this used to always set the animation to "move" and return true (= using non-idle animation).

So it never felt back to idle.

I think this committed code is the intended behavior and my quick test seems to indicate so.